move bulk monitors hook to monitorHooks
